fix(chat): handle malformed mermaid data urls
