+++ packages/vscode/vite.config.ts
@@ -47,7 +47,7 @@ export default defineConfig(({ mode }) => ({
     },
   },
   optimizeDeps: {
-    include: ['@opencode-ai/sdk/v2', 'shiki'],
+    include: ['@opencode-ai/sdk/v2'],
   },
   build: {
     outDir: path.resolve(__dirname, 'dist/webview'),
